# cmd/docs/swagger.yaml
definitions:
...
  dto.CreateUserRequestDTO:
    properties:
      email:
        type: string
      isAdmin:
        type: boolean
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  dto.LoginDTO:
    properties:
      email:
        example: user@email.com
        type: string
      password:
        example: pass1234
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponseDTO:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
...
  httputil.HttpError400:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  httputil.HttpError401:
    properties:
      code:
        example: 401
        type: integer
      message:
        example: Unauthorized
        type: string
    type: object
  httputil.HttpError500:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: internal server error
        type: string
    type: object
...
info:
  contact:
    email: sankaungkin@gmail.com
    name: API Support
  description: This is an auto-generated API docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login Data
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      summary: Login to the api with email and password
      tags:
      - Auth
  /api/auth/logout:
    post:
      description: Logout user
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      summary: Logout user
      tags:
      - Auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Get refresh token
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      summary: Get refresh token
      tags:
      - Auth
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: Create new user based on parameters
      parameters:
      - description: Signup Data
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequestDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      summary: Create new user based on parameters
      tags:
      - Auth
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"